// UI/student-affairs/src/app/models/student-subject.model.ts
export interface StudentSubject {
     id: number;
    student_ID: number;
    student_Name: string;
    subject_ID: number;
    subject_Name: string;
  }

  export interface StudentSubjectList {
  ID: number;
   Student_ID: number;
   Student_Name: string;
   Subject_ID: number;
   Subject_Name: string;
 }